KAN-21 Wait for user to click on Next after the last question
The previous behavior was after user click 'Check' we display the correct / wrong count.
This would cause confusion to users because they don't actually know if the last question was correct or not.

We're using a user interaction now to trigger the summary screen so user can see if they are correct for the last question, and what the correct note is should they were wrong.